+++ cmd/staking-api-service/main.go
@@ -7,7 +7,6 @@ import (
 	"github.com/babylonlabs-io/staking-api-service/cmd/staking-api-service/cli"
 	"github.com/babylonlabs-io/staking-api-service/cmd/staking-api-service/scripts"
 	"github.com/babylonlabs-io/staking-api-service/internal/shared/api"
-	"github.com/babylonlabs-io/staking-api-service/internal/shared/api/services"
 	"github.com/babylonlabs-io/staking-api-service/internal/shared/config"
 	dbclients "github.com/babylonlabs-io/staking-api-service/internal/shared/db/clients"
 	dbmodel "github.com/babylonlabs-io/staking-api-service/internal/shared/db/model"
@@ -15,6 +14,7 @@ import (
 	"github.com/babylonlabs-io/staking-api-service/internal/shared/observability/healthcheck"
 	"github.com/babylonlabs-io/staking-api-service/internal/shared/observability/metrics"
 	queueclients "github.com/babylonlabs-io/staking-api-service/internal/shared/queue/clients"
+	"github.com/babylonlabs-io/staking-api-service/internal/shared/services"
 	"github.com/babylonlabs-io/staking-api-service/internal/shared/types"
 	"github.com/joho/godotenv"
 	"github.com/rs/zerolog/log"
