+++ internal/shared/queue/handlers/handlers.go
@@ -1,8 +1,8 @@
 package queuehandlers
 
 import (
-	"github.com/babylonlabs-io/staking-api-service/internal/shared/api/services"
 	queuehandler "github.com/babylonlabs-io/staking-api-service/internal/shared/queue/handler"
+	"github.com/babylonlabs-io/staking-api-service/internal/shared/services"
 	v1queuehandler "github.com/babylonlabs-io/staking-api-service/internal/v1/queue/handler"
 	v2queuehandler "github.com/babylonlabs-io/staking-api-service/internal/v2/queue/handler"
 )
