+++ internal/shared/services/service/interface.go
