+++ internal/v1/api/handlers/handler.go
@@ -3,8 +3,8 @@ package v1handlers
 import (
 	"context"
 
-	handler "github.com/babylonlabs-io/staking-api-service/internal/shared/api/handler"
-	v1service "github.com/babylonlabs-io/staking-api-service/internal/v1/api/service"
+	handler "github.com/babylonlabs-io/staking-api-service/internal/shared/api/handlers/handler"
+	v1service "github.com/babylonlabs-io/staking-api-service/internal/v1/service"
 )
 
 type V1Handler struct {
