+++ internal/v1/api/handlers/staker.go
@@ -3,7 +3,7 @@ package v1handlers
 import (
 	"net/http"
 
-	"github.com/babylonlabs-io/staking-api-service/internal/shared/api/handler"
+	"github.com/babylonlabs-io/staking-api-service/internal/shared/api/handlers/handler"
 	"github.com/babylonlabs-io/staking-api-service/internal/shared/types"
 	"github.com/babylonlabs-io/staking-api-service/internal/shared/utils"
 )
