+++ internal/v1/api/handlers/stats.go
@@ -3,9 +3,9 @@ package v1handlers
 import (
 	"net/http"
 
-	"github.com/babylonlabs-io/staking-api-service/internal/shared/api/handler"
+	"github.com/babylonlabs-io/staking-api-service/internal/shared/api/handlers/handler"
 	"github.com/babylonlabs-io/staking-api-service/internal/shared/types"
-	v1service "github.com/babylonlabs-io/staking-api-service/internal/v1/api/service"
+	v1service "github.com/babylonlabs-io/staking-api-service/internal/v1/service"
 )
 
 // GetOverallStats gets overall stats for babylon staking
