+++ internal/v1/service/service.go
@@ -1,16 +1,16 @@
-package v2service
+package v1service
 
 import (
 	"context"
 
-	service "github.com/babylonlabs-io/staking-api-service/internal/shared/api/service"
+	"github.com/babylonlabs-io/staking-api-service/internal/shared/services/service"
 	"github.com/babylonlabs-io/staking-api-service/internal/shared/config"
 	dbclients "github.com/babylonlabs-io/staking-api-service/internal/shared/db/clients"
 	"github.com/babylonlabs-io/staking-api-service/internal/shared/http/clients"
 	"github.com/babylonlabs-io/staking-api-service/internal/shared/types"
 )
 
-type V2Service struct {
+type V1Service struct {
 	*service.Service
 }
 
@@ -21,13 +21,13 @@ func New(
 	finalityProviders []types.FinalityProviderDetails,
 	clients *clients.Clients,
 	dbClients *dbclients.DbClients,
-) (*V2Service, error) {
+) (*V1Service, error) {
 	service, err := service.New(ctx, cfg, globalParams, finalityProviders, clients, dbClients)
 	if err != nil {
 		return nil, err
 	}
 
-	return &V2Service{
+	return &V1Service{
 		service,
 	}, nil
 }
