+++ tests/integration_test/finality_provider_test.go
@@ -9,13 +9,13 @@ import (
 	"testing"
 	"time"
 
-	handler "github.com/babylonlabs-io/staking-api-service/internal/shared/api/handler"
+	handler "github.com/babylonlabs-io/staking-api-service/internal/shared/api/handlers/handler"
 	"github.com/babylonlabs-io/staking-api-service/internal/shared/config"
 	"github.com/babylonlabs-io/staking-api-service/internal/shared/db"
 	dbclients "github.com/babylonlabs-io/staking-api-service/internal/shared/db/clients"
 	"github.com/babylonlabs-io/staking-api-service/internal/shared/types"
-	v1service "github.com/babylonlabs-io/staking-api-service/internal/v1/api/service"
 	v1dbmodel "github.com/babylonlabs-io/staking-api-service/internal/v1/db/model"
+	v1service "github.com/babylonlabs-io/staking-api-service/internal/v1/service"
 	testmock "github.com/babylonlabs-io/staking-api-service/tests/mocks"
 	"github.com/babylonlabs-io/staking-api-service/tests/testutils"
 	"github.com/stretchr/testify/assert"
