+++ tests/integration_test/unbonding_test.go
@@ -10,14 +10,14 @@ import (
 	"time"
 
 	"github.com/babylonlabs-io/staking-api-service/internal/shared/api"
-	handler "github.com/babylonlabs-io/staking-api-service/internal/shared/api/handler"
+	handler "github.com/babylonlabs-io/staking-api-service/internal/shared/api/handlers/handler"
 	"github.com/babylonlabs-io/staking-api-service/internal/shared/config"
 	dbclients "github.com/babylonlabs-io/staking-api-service/internal/shared/db/clients"
 	dbmodel "github.com/babylonlabs-io/staking-api-service/internal/shared/db/model"
 	"github.com/babylonlabs-io/staking-api-service/internal/shared/types"
 	v1handlers "github.com/babylonlabs-io/staking-api-service/internal/v1/api/handlers"
-	v1service "github.com/babylonlabs-io/staking-api-service/internal/v1/api/service"
 	v1dbmodel "github.com/babylonlabs-io/staking-api-service/internal/v1/db/model"
+	v1service "github.com/babylonlabs-io/staking-api-service/internal/v1/service"
 	testmock "github.com/babylonlabs-io/staking-api-service/tests/mocks"
 	"github.com/babylonlabs-io/staking-api-service/tests/testutils"
 	"github.com/babylonlabs-io/staking-queue-client/client"
